Removed trash (add) icon from unselected
Before we had the trash icon for the classes not selected which served
as the add button, but it broke once we switched to multiple droppables.
I just removed the icon for this since it doesn't really make sense to
have an add button if a class corresponds with more than one quarter.
Old code commented out